+++ src/main/java/floud/demo/common/response/Error.java
@@ -1,4 +1,4 @@
-package floud.demo.common;
+package floud.demo.common.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -7,12 +7,14 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum Success {
+public enum Error {
     // Default
-    SUCCESS(HttpStatus.OK, "Request successfully processed"),
+    ERROR(HttpStatus.BAD_REQUEST, "Request processing failed"),
 
+
+    // 500 INTERNAL SERVER ERROR
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error"),
     ;
-
 
     private final HttpStatus httpStatus;
     private final String message;
